Rename core to kernel, bc rust core is also name core and it's cumbersome to qualify with crate
