+++ src/kernel/extract.rs
